+++ tests/uri.js
@@ -7,14 +7,20 @@ describe('URI', function() {
     var location = URI('?a=b');
 
     location.should.have.property('query');
+    location.should.have.property('search');
+
     location.query.should.eql({ a: 'b' });
+    location.search.should.eql('?a=b');
   });
 
   it('should work with full uris', function() {
     var location = URI('https://user:pw@example.com:80/a/b?c=d#e');
 
     location.should.have.property('query');
+    location.should.have.property('search');
+
     location.query.should.eql({ c: 'd' });
+    location.search.should.eql('?c=d');
   });
 
   describe('query', function() {
@@ -59,6 +65,15 @@ describe('URI', function() {
       parsed.should.eql({ c: 'd' });
     });
     
+  });
+
+  describe('extract_query', function() {
+
+    it('should work with full urls', function() {
+      var query = URI.extract_query('https://user:pw@example.com:80/a/b?c=d&e=f&g#e');
+
+      query.should.eql('?c=d&e=f&g');
+    });
   });
 
   describe('extend', function() {