+++ Changelog.md
@@ -8,6 +8,7 @@
  - Add `uriCallback` method of getting `URI`
  - Passing full URLs to URI.query
  - Getting Location-like objects by simply calling `URI`. 
+ - Add a search parameter to the Location-like object.
 
 ## Version 0.1.1